ras_orchestrator/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
redis==5.0.1
kafka-python==2.0.3
pyyaml==6.0.1
prometheus-client==0.19.0
python-json-logger==2.0.7
